Change balances to decimal. Change inputs to decimal. Fix bug with Period

+++ demo-app/src/SendmeDemo.API.Host/Contracts/IERC20.cs
@@ -2,11 +2,11 @@ using SendmeDemo;
 
 public interface IERC20
 {
-    Task<int> GetTotalSupplyAsync();
-    Task<string> TransferAsync(Wallet from, string to, int value);
+    Task<decimal> GetTotalSupplyAsync();
+    Task<string> TransferAsync(Wallet from, string to, decimal value);
     Task SetPeriodAsync(Wallet wallet, int period);
     Task SetLimitAsync(Wallet wallet, int limit);
-    Task<string> MintAsync(Wallet issuer, string to, int value);
-    Task<string> BurnAsync(Wallet issuer, int value);
-    Task<int> GetBalanceAsync(string address);
+    Task<string> MintAsync(Wallet issuer, string to, decimal value);
+    Task<string> BurnAsync(Wallet issuer, decimal value);
+    Task<decimal> GetBalanceAsync(string address);
 }
