+++ demo-app/src/SendmeDemo.API.Host/Models/UserModel.cs
@@ -5,7 +5,7 @@ public class UserModel
     public UserModel()
     { }
     
-    public UserModel(string name, string address, List<string> properties, int balance)
+    public UserModel(string name, string address, List<string> properties, decimal balance)
     {
         Name = name;
         Address = address;
@@ -16,12 +16,12 @@ public class UserModel
     public string Name { get; set; }
     public string Address { get; set; }
     public List<string> Properties { get; set; }
-    public int Balance { get; set; }
+    public decimal Balance { get; set; }
 }
 
 public class IssuerModel : UserModel
 {
-    public IssuerModel(UserModel model, int totalSupply)
+    public IssuerModel(UserModel model, decimal totalSupply)
     {
         Name = model.Name;
         Address = model.Address;
@@ -29,5 +29,5 @@ public class IssuerModel : UserModel
         Balance = model.Balance;
         TotalSupply = totalSupply;
     }
-    public int TotalSupply { get; set; }
+    public decimal TotalSupply { get; set; }
 }
